+++ lib/commands/compareElement.js
@@ -1,6 +1,6 @@
-import { executeCompare, instanceCompareOptions } from '../helpers/compare';
-import { checkImageExists } from '../helpers/image';
-import { formatFileName } from '../helpers/utils';
+import { instanceCompareOptions } from '../helpers/options';
+import { checkBaselineImageExists } from '../methods/checkBaselineImageExists';
+import { executeCompare } from '../methods/executeCompare';
 
 /**
  * Compare an element against a baseline
@@ -19,6 +19,7 @@ import { formatFileName } from '../helpers/utils';
  * @example
  *  // Default
  *  browser.compareElement('~selector', 'tag-name-of-image');
+ *
  *  // With compare options
  *  browser.compareElement(
  *      '~selector',
@@ -31,6 +32,33 @@ import { formatFileName } from '../helpers/utils';
  *      }
  *  );
  *
+ *  // With all resize dimensions and ignore alpha
+ *  browser.saveElement(
+ *      '~selector',
+ *      'tag-name-of-image',
+ *      {
+ *          ignoreAlpha: true,
+ *          resizeDimensions: {
+ *              top: 10,
+ *              right: 20,
+ *              bottom: 30,
+ *              left: 40,
+ *          }
+ *      }
+ *  );
+ *
+ *  // With only 1 resize dimension and ignore alpha
+ *  browser.saveElement(
+ *      '~selector',
+ *      'tag-name-of-image',
+ *      {
+ *          ignoreAlpha: true,
+ *          resizeDimensions: {
+ *              left: 10,
+ *          }
+ *      }
+ * );
+ *
  * @return {Promise<{
  *      fileName: {string},
  *      folders: {object},
@@ -38,21 +66,26 @@ import { formatFileName } from '../helpers/utils';
  * }>}
  */
 export default async function compareElement(element, tag, options = {}) {
+    // Determine the element compare options
     const elementCompareOptions = {
         ...(this.compareOptions),
         ...(instanceCompareOptions(options)),
     };
-    const fileName = formatFileName(this.formatString, tag);
-    await browser.saveElement(element, tag, options);
-    await checkImageExists(
+
+    // Save the element and return the data
+    const elementScreenshotData = await browser.saveElement(element, tag, options);
+
+    // Check if the baseline image exists
+    await checkBaselineImageExists(
         this.folders,
-        fileName,
+        elementScreenshotData.fileName,
         this.autoSaveBaseline,
     );
 
+    // Execute the compare and return the compare data
     return await executeCompare(
         this.folders,
-        fileName,
+        elementScreenshotData.fileName,
         elementCompareOptions,
     );
 }
