+++ lib/commands/compareScreen.js
@@ -1,7 +1,7 @@
-import { executeCompare, instanceCompareOptions } from '../helpers/compare';
-import { checkImageExists } from '../helpers/image';
-import { determineIgnoreRectangles } from '../helpers/rectangles';
-import { formatFileName } from '../helpers/utils';
+import { instanceCompareOptions } from '../helpers/options';
+import { checkBaselineImageExists } from '../methods/checkBaselineImageExists';
+import { determineIgnoreRectangles } from '../methods/determineIgnoreRectangles';
+import { executeCompare } from '../methods/executeCompare';
 
 /**
  * Compare a screen against a baseline
@@ -31,27 +31,36 @@ import { formatFileName } from '../helpers/utils';
  *      }
  *  );
  *
- * @return {Promise<number>}
+ * @return {Promise<{
+ *      fileName: {string},
+ *      folders: {object},
+ *      misMatchPercentage: number
+ * }>}
  */
 export default async function compareScreen(tag, options = {}) {
+    // Determine the screen compare options
     const screenCompareOptions = {
         ...(this.compareOptions),
         ...(instanceCompareOptions(options)),
     };
-    const fileName = formatFileName(this.formatString, tag);
-    const bufferedScreenshot = await browser.saveScreen(tag, true);
 
-    await checkImageExists(
+    // Save the screen and return the data
+    const screenshotData = await browser.saveScreen(tag, true);
+
+    // Check if the baseline image exists
+    await checkBaselineImageExists(
         this.folders,
-        fileName,
+        screenshotData.fileName,
         this.autoSaveBaseline,
     );
 
-    screenCompareOptions.ignoreRectangles = await determineIgnoreRectangles(bufferedScreenshot, screenCompareOptions);
+    // Determine the ignore rectangles
+    screenCompareOptions.ignoreRectangles = await determineIgnoreRectangles(screenshotData.base64Screenshot, screenCompareOptions);
 
+    // Execute the compare and return the compare data
     return await executeCompare(
         this.folders,
-        fileName,
+        screenshotData.fileName,
         screenCompareOptions,
     );
 }
