+++ lib/commands/saveScreen.js
@@ -1,30 +1,39 @@
 import { resolve } from 'path';
 import { takeBase64Screenshot } from '../helpers/takeBase64Screenshot';
 import { saveImage } from '../helpers/image';
-import { formatFileName } from '../helpers/utils';
+import { determineFileData, formatFileName } from '../helpers/utils';
+import saveBase64Image from '../methods/saveBase64Image';
 
 /**
  * Save a screen
  *
- * @param {string} tag
- * @param {boolean} returnBuffer
+ * @param {string} tag The tag of the image name
  *
  * @example
  *  browser.saveScreen('tag-name-of-image');
  *
- * @return {Promise<buffer|object>}
+ * @return {Promise<{
+ *  base64Screenshot: string,
+ *  fileName: string,
+ *  folders: {
+ *      actual: string,
+ *  },
+ * }>}
  */
-export default async function saveScreen(tag, returnBuffer = false) {
-    // Determine filename
-    const fileName = formatFileName(this.formatString, tag);
-    const filePath = resolve(this.folders.actual, fileName);
+export default async function saveScreen(tag) {
+    // Determine file data
+    const fileData = determineFileData(this.folders.actual, this.formatString, tag);
+
     // Take the screenshot
-    const buffer = await takeBase64Screenshot();
+    const base64Screenshot = await takeBase64Screenshot();
 
-    saveImage(filePath, buffer);
+    // Save the screenshot
+    await saveBase64Image(fileData.filePath, base64Screenshot);
 
-    return returnBuffer ? buffer : {
-        fileName,
+    // Return the screenshot data
+    return {
+        base64Screenshot,
+        fileName: fileData.fileName,
         folders:{
             actual: this.folders.actual
         }
