+++ lib/helpers/constants.js
@@ -1,19 +1,25 @@
 export const ACTUAL = 'actual';
 export const DEFAULT_IMAGE_FORMAT_NAME = '{tag}-{deviceName}';
 export const DEFAULT_BASELINE_FODLER = './baseline';
+export const DEFAULT_RESIZE_DIMENSIONS = {
+    top: 0,
+    right: 0,
+    bottom: 0,
+    left: 0
+};
 export const DEFAULT_SCREENSHOTS = './.tmp/screenshots';
 export const DIFF = 'diff';
-export const NO_DEVICE_NAME = 'no-device-name-in-caps';
-export const STATUS_BAR_SELECTORS = {
-    IOS: '//XCUIElementTypeStatusBar',
-    ANDROID: '//*[@resource-id="android:id/statusBarBackground"]'
-};
-export const NAVIGATION_BAR_SELECTORS = {
-    ANDROID: '//*[@resource-id="android:id/navigationBarBackground"]'
-};
 export const IPHONE_X_BOTTOM_BAR = {
     height: 16,
     width: 402,
     x: 362,
     y: 2396
+};
+export const NAVIGATION_BAR_SELECTORS = {
+    ANDROID: '//*[@resource-id="android:id/navigationBarBackground"]'
+};
+export const NO_DEVICE_NAME = 'no-device-name-in-caps';
+export const STATUS_BAR_SELECTORS = {
+    IOS: '//XCUIElementTypeStatusBar',
+    ANDROID: '//*[@resource-id="android:id/statusBarBackground"]'
 };
