+++ lib/helpers/options.js
@@ -1,59 +1,3 @@
-import { readFileSync } from 'fs-extra';
-import { join, resolve } from 'path';
-import { compareImages } from '../resemble/compareImages';
-import { saveImage } from './image';
-
-/**
- * Compare 2 images with each other
- *
- * @param {object}  folders
- * @param {string}  fileName
- * @param {object}  options
- *
- * @return {Promise<{
- *      fileName: {string},
- *      folders: {object},
- *      misMatchPercentage: number
- * }>}
- */
-export async function executeCompare(folders, fileName, options) {
-    if (options.ignoreAlpha) {
-        options.ignore.push('alpha')
-    }
-    if (options.ignoreAntialiasing) {
-        options.ignore.push('antialiasing')
-    }
-    if (options.ignoreColors) {
-        options.ignore.push('colors')
-    }
-    if (options.ignoreLess) {
-        options.ignore.push('less')
-    }
-    if (options.ignoreNothing) {
-        options.ignore.push('nothing')
-    }
-
-    // Execute the compare and retrieve the data
-    const data = await compareImages(
-        readFileSync(resolve(folders.baseline, fileName)),
-        readFileSync(resolve(folders.actual, fileName)),
-        options
-    );
-    const misMatchPercentage = options.rawMisMatchPercentage
-        ? data.rawMisMatchPercentage
-        : Number(data.rawMisMatchPercentage.toFixed(2));
-
-    if (misMatchPercentage > options.saveAboveTolerance || options.debug) {
-        saveImage(resolve(folders.diff, fileName), data.getBuffer());
-    }
-
-    return {
-        fileName,
-        folders,
-        misMatchPercentage,
-    };
-}
-
 /**
  * Determine the default compare options
  *
