+++ lib/helpers/utils.js
@@ -2,8 +2,6 @@ import { ensureDirSync } from 'fs-extra';
 import { resolve } from 'path';
 import { NO_DEVICE_NAME } from './constants';
 
-let DEVICE_INFO = null;
-
 /**
  * Create the filename
  *
@@ -49,37 +47,26 @@ export function determineFolder(folder, savePerDevice) {
 }
 
 /**
- * Get the device info and store it to a global if needed
+ * Determine the filename and file path
  *
- * @param   {string}  screenshot
+ * @param {string}  folder
+ * @param {string}  formatString
+ * @param {string}  tag
  *
- * @return  {Promise<{
- *      dpr: number,
- *      screensize: {
- *          height: number,
- *          width: number,
- *      },
- *      screenshotHeight: number,
- *      screenshotWidth: number,
- *      isIphoneX: boolean,
- * }>}
+ * @return {
+ *  {
+ *      fileName: string,
+ *      filePath: string,
+ *  }
+ * }
  */
-export async function getDeviceInfo(screenshot) {
-    if (DEVICE_INFO) {
-        return DEVICE_INFO;
-    }
+export function determineFileData(folder, formatString, tag){
+    const fileName = formatFileName(formatString, tag);
 
-    const screenshotWidth = new Buffer(screenshot, 'base64').readUInt32BE(16);
-    const screensize = (await browser.windowHandleSize()).value;
-    DEVICE_INFO = {
-        dpr: screenshotWidth / screensize.width,
-        screensize,
-        screenshotHeight: new Buffer(screenshot, 'base64').readUInt32BE(20),
-        screenshotWidth,
-        isIphoneX: (screensize.width === 812 || screensize.height === 812) && browser.isIOS,
+    return {
+        fileName,
+        filePath: resolve(folder, fileName),
     };
-
-    return DEVICE_INFO;
 }
 
 /**
