+++ lib/index.js
@@ -1,6 +1,5 @@
 import { join, normalize, resolve } from 'path';
 import { ensureDirSync, writeFileSync } from 'fs-extra';
-import { defaultCompareOptions } from './helpers/compare';
 import {
     ACTUAL,
     DEFAULT_BASELINE_FODLER,
@@ -12,6 +11,7 @@ import compareScreen from './commands/compareScreen';
 import compareElement from './commands/compareElement';
 import saveElement from './commands/saveElement';
 import saveScreen from './commands/saveScreen';
+import { defaultCompareOptions } from './helpers/options';
 import { determineFolder } from './helpers/utils';
 
 /**
