+++ lib/methods/determineIgnoreRectangles.js
@@ -1,6 +1,7 @@
-import { IPHONE_X_BOTTOM_BAR, NAVIGATION_BAR_SELECTORS, STATUS_BAR_SELECTORS } from './constants';
-import { getDeviceInfo, multiplyIosObjectValuesAgainstDPR } from './utils';
+import { IPHONE_X_BOTTOM_BAR, NAVIGATION_BAR_SELECTORS, STATUS_BAR_SELECTORS } from '../helpers/constants';
+import { multiplyIosObjectValuesAgainstDPR } from '../helpers/utils';
 import { red, yellow } from 'chalk';
+import { getDeviceInfo } from './getDeviceInfo';
 
 let NAVIGATION_BAR_DIMENSIONS = null;
 let STATUS_BAR_DIMENSIONS = null;
@@ -8,15 +9,15 @@ let STATUS_BAR_DIMENSIONS = null;
 /**
  * Determine the ignore rectangles
  *
- * @param {buffer}  bufferedScreenshot
+ * @param {string}  base64Image
  * @param {object}  options
  *
  * @return {Promise<Array>}
  */
-export async function determineIgnoreRectangles(bufferedScreenshot, options) {
+export async function determineIgnoreRectangles(base64Image, options) {
     let rectangles = [];
     let allowInvisibleElements;
-    const deviceInfo = await getDeviceInfo(bufferedScreenshot);
+    const deviceInfo = await getDeviceInfo(base64Image);
 
     // With Android and UiAutomator2 the status bar is not in the view by default
     // It can be seen when {settings: {allowInvisibleElements: true}} is set
