+++ lib/methods/makeCroppedBase64Image.js
@@ -1,12 +1,11 @@
 import { yellow } from 'chalk';
 import { createWriteStream } from 'fs-extra';
+import { DEFAULT_RESIZE_DIMENSIONS } from '../helpers/constants';
 
 const Canvas = require('canvas-prebuilt');
 
 /**
  * Save a (cropped) canvas screen
- *
- * @TODO: also check right and bottom coordinates
  *
  * @param {buffer}  screenshot
  * @param {object}  rectangles
@@ -15,28 +14,26 @@ const Canvas = require('canvas-prebuilt');
 export function makeCroppedBase64Image(screenshot, rectangles, resizeDimensions) {
     const CanvasImage = Canvas.Image;
     const image = new CanvasImage();
+    const resize = {...DEFAULT_RESIZE_DIMENSIONS, ...resizeDimensions};
     let canvas;
-
-    console.log('rectangles = ', rectangles);
-    console.log('resizeDimensions = ', resizeDimensions);
 
     image.onerror = (err) => {
         throw err
     };
 
     image.onload = () => {
-        const w = rectangles.width + resizeDimensions.left + resizeDimensions.right;
-        const h = rectangles.height + resizeDimensions.top + resizeDimensions.bottom;
+        const w = rectangles.width + resize.left + resize.right;
+        const h = rectangles.height + resize.top + resize.bottom;
         canvas = new Canvas(w, h);
         const ctx = canvas.getContext('2d');
 
-        let sourceXStart = rectangles.x - resizeDimensions.left;
-        let sourceYStart = rectangles.y - resizeDimensions.top;
+        let sourceXStart = rectangles.x - resize.left;
+        let sourceYStart = rectangles.y - resize.top;
 
         if (sourceXStart < 0) {
             console.log(yellow(`
 #####################################################################################
- THE RESIZE DIMENSION LEFT '${resizeDimensions.left}' MADE THE CROPPING GO OUT OF
+ THE RESIZE DIMENSION LEFT '${resize.left}' MADE THE CROPPING GO OUT OF
  THE IMAGE BOUNDARIES RESULTING IN AN IMAGE STARTPOSITION '${sourceXStart}'.
  THIS HAS BEEN DEFAULTED TO '0'
 #####################################################################################
@@ -47,7 +44,7 @@ export function makeCroppedBase64Image(screenshot, rectangles, resizeDimensions)
         if (sourceYStart < 0) {
             console.log(yellow(`
 #####################################################################################
- THE RESIZE DIMENSION LEFT '${resizeDimensions.top}' MADE THE CROPPING GO OUT OF
+ THE RESIZE DIMENSION LEFT '${resize.top}' MADE THE CROPPING GO OUT OF
  THE IMAGE BOUNDARIES RESULTING IN AN IMAGE STARTPOSITION '${sourceYStart}'.
  THIS HAS BEEN DEFAULTED TO '0'
 #####################################################################################
