+++ lib/methods/saveBase64Image.js
@@ -3,11 +3,11 @@ import { outputFile } from 'fs-extra';
 /**
  * Save the base64 image to a file
  *
- * @param   {string}  base64Image
  * @param   {string}  filePath
+ * @param   {string}  base64Image
  *
  * @returns {Promise<void>}
  */
-export default async function saveBase64Image(base64Image, filePath) {
+export default async function saveBase64Image(filePath, base64Image) {
     return outputFile(filePath, base64Image, 'base64');
 }
