+++ lib/methods/takeBase64Screenshot.js
@@ -1,5 +1,5 @@
 /**
- * Take a screenshot
+ * Take a base64 screenshot
  *
  * @return {Promise<string>}
  */
